Depth-first, cross-component iterator.
A Cross-component, depth-first iterator, that iterates
through a whole graph, traversing edges only following
their direction.

With A -> B, the iterator will move from A to B, but
not from B to A.

This iterator is a cross-component iterator, meaning that when
the iteration through a connected-component of the graph is
finished, the iterator jumps to the next component automatically.
This ensures that all the vertices of the graph are iterated
exactly once with this iterator.

The order in which the connected-components are iterated can be
specified by using the constructor that specifies an iterable
over the collection of roots, using a list with the desired
order. For this iterator to operate properly and indeed iterate
through all the vertices of the graph exactly once, the specified
collection of roots must include all the roots of the graph, that
is: all the vertices that have no incoming edges.

Within a single connected-component, the order of iteration is
depth-first. The iterator only jumps to another component only
when all the vertices of the currently iterated component have
been iterated.