Fix uses of ref objects in BranchGraph#checkFuse.
Interestingly, it does not fix the test from the previous commit,
but makes it fail in a proper way. Since we cannot add an edge
between 2 branch vertices that are already linked by an edge,
the addEdge() method returns (correctly) null. And the test
generates a NPE.

Previously the uninitialized reference was not null and the test
was failing without throwing a NPE.