Support non-simple, directed graphs based on Pools.
So far, all Graphs based on AbstractEdgePool were simple, directed
graphs. For these graphs, there is at most one directed
edge between a source and a target vertices (0 or 1, but
there might another edge in the opposite direction). This
is enforced here by preventing adding an edge between a
source and a target already linked by an edge.

This commit inserts an AbstractNonSimpleEdgePool in the hierarchy
(as the mother class of AbstractEdgePool) that does not have
this restriction.