Add a JUnit test that fails over edges iteration.
After all edges have been removed from a vertex, the edge collection
(incoming, outgoing or allEdges) have all the proper size, but the
iterator behaves weirdly. It indeed correctly report hasNext=false,
but when prompted for .next(), it returns a non-null edge instance
which is incorrectly instantiated.
The iterator contract states that we should throw a NoSuchElement
exception.