Alternate solution for branch graph: use an external graph listener.
The listener takes a ListenableGraph and a non-simple graph. It
builds the branch graph on the non-simple graph using events
triggered on the listenable graph.

This way, we decouple the branch graph logic from the data structures
that stores it.

This is not perfect yet, as we miss a way to clear that structure when
the graph is rebuilt.