Generate a map of 'tracks' from a graph.
A track is a connected component of the graph. Here, the connected
components are exposed via their collection of vertices or edges,
and can be accessed via a map where a root vertex is used as a key.

A track map is only a 'snapshot' of the graph, and is not kept in
sync with it if it is modified.

It is ensured that each vertex and edge is present exactly once
in the track maps. Even if a track has more than 1 root, it is
present only once in the maps. Since we use roots to build and
index the maps, connected components that have no roots (loops)
will not be included in the map.