Fix one bug in the BranchGraphImp class.
When splitting a branch edge we were assuming that the core vertex
we split the branch on has only one outgoing edge. Of course this
is false in the case the branch graph is rebuilt on a core graph that
already has many divisions.

This commit fixes the issue by actually searching for the right
outgoing edge of the specified vertex to split. Beyond the verbose
comments I had to add to keep my sanity, the important lines are:

```java
E outgoingEdge = null;
for ( final E e : v.outgoingEdges() )
{
	final BE be = ebeMap.get( e );
	if ( be != null && be.equals( initialBE ) )
	{
		outgoingEdge = e;
		break;
	}
}
```

instead of:

``` java
final E outgoingEdge = v.outgoingEdges().get( 0, refE1 );
```