Brutal way to solve the bug with linkBranchEdge method.
We will not assume that the vertices of the core graph we iterate
have exactly one outgoing edge. Instead we will search for a path
from begin to start and map the two graphs on the path found only.
Hopefully the bug that caused hanging when creating the branch graph
is fixed by this now.