Do not register the branch graph as a listener of the core graph.
Currently we have problems in the Mamut app when the user makes a
batch deletion while the branch graph is listening to the core graph.
It might be an issue with a missing reading lock.
These errors are tough to corner.
In the meantime, we request that the branch graph is either
- explicitely added as a listener to the core graph when safe, or
- 'manually' refreshed by the consumer by calling the graphRebuilt()
method which is efficient.