Refactor the branch-graph.
There are now two abstract branch-graph implementations.

The BranchGraphImp class does not implement incremental changes.
The branch-graph is only rebuilt upon calls of the graphRebuilt()
method. The methods edgeAdded(Edge), edgeRemoved(Edge),
vertexAdded(Vertex) and vertexRemoved(Vertex) do not do anything.
As a side effect, this implementation does not support 'rings'.
If aconnected component in the core-graph looks like this:
	 A -> B -> C -> D -> A
(edge direction is important), it won't be discovered in the
branch-graph,which requires 'roots' to be built. Roots are
vertices with no incomingedges. 'Diamonds' connected components:
	A -> B -> D
and
	A -> C -> D
are supported.

The BranchGraphImpIncremental class supports incremental changes.
But,  But there are still bugs when making batch changes in the
core-graph. They are probably linked to not locking the core-graph
and not cleaning the maps. This class should not be used without
uttermost cautious, and could do with a revamp.

There are JUnit tests for both implementations.

There a lot of common codes between the two implementations but we
cannot subclass one with the other because we need to specifically
call some methods in the superclass.

This is a crude way of casting away a piece of code I do not trust.